security/requirements.txt
fastapi
python-multipart
uvicorn[standard]
python-jose[crytography]
passlib[bcrypt]
psycopg2
requests